+++ utils/misc/create_iudx_type.py
@@ -11,16 +11,22 @@ for filename in glob.glob(os.path.join(folder_path, '*.jsonld')):
     with open(filename, "r+") as obj_file:
         new_dict = OrderedDict()
         obj = json.load(obj_file)
-        new_dict["@context"] = obj["@context"]
-        del(obj["@context"]) 
-        new_dict["@graph"] = obj["@graph"]
-        for field in new_dict["@graph"]:
-            if isinstance(field["@type"], list):
-                iudx_type = [i for i in field["@type"] if "iudx:" in i]
-                if len(iudx_type) == 1:
-                    field["@type"] = str(iudx_type[0])
-                else:
-                    print("Error: Number of @type properties greater than 2")
-        obj_file.seek(0)
-        json.dump(new_dict, obj_file, indent=4)
-        obj_file.truncate()
+        if "@context" in obj.keys():
+            new_dict["@context"] = obj["@context"]
+            del(obj["@context"]) 
+        else:
+            print("@context missing in " + filename)
+        if "@graph" in obj.keys():
+            new_dict["@graph"] = obj["@graph"]
+            for field in new_dict["@graph"]:
+                if isinstance(field["@type"], list):
+                    iudx_type = [i for i in field["@type"] if "iudx:" in i]
+                    if len(iudx_type) == 1:
+                        field["@type"] = str(iudx_type[0])
+                    else:
+                        print("Error: Number of @type properties greater than 2")
+            obj_file.seek(0)
+            json.dump(new_dict, obj_file, indent=4)
+            obj_file.truncate()
+        else:
+            print("@graph missing in " + filename)
